chore: fix warnings
lvgl/src/core/lv_refr.c: In function 'call_flush_cb':
lvgl/src/core/lv_refr.c:1239:17: warning: format '%d' expects argument of type 'int', but argument 6 has type 'lv_coord_t {aka const long int}' [-Wformat=]
     REFR_TRACE("Calling flush_cb on (%d;%d)(%d;%d) area with %p image pointer", area->x1, area->y1, area->x2, area->y2,
                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~  ~~~~~~~~
lvgl/src/core/lv_refr.c:1239:17: warning: format '%d' expects argument of type 'int', but argument 7 has type 'lv_coord_t {aka const long int}' [-Wformat=]
lvgl/src/core/lv_refr.c:1239:17: warning: format '%d' expects argument of type 'int', but argument 8 has type 'lv_coord_t {aka const long int}' [-Wformat=]
lvgl/src/core/lv_refr.c:1239:17: warning: format '%d' expects argument of type 'int', but argument 9 has type 'lv_coord_t {aka const long int}' [-Wformat=]

lvgl/src/misc/lv_timer.c: In function 'lv_timer_handler':
lvgl/src/misc/lv_timer.c:144:90: warning: format '%d' expects argument of type 'int', but argument 6 has type 'uint32_t {aka long unsigned int}' [-Wformat=]
     TIMER_TRACE("finished (%d ms until the next timer call)", time_till_next);
                                                                                          ^

lvgl/src/widgets/roller/lv_roller.c: In function 'lv_roller_set_options':
lvgl/src/widgets/roller/lv_roller.c:127:105: warning: format '%d' expects argument of type 'int', but argument 6 has type 'uint32_t {aka long unsigned int}' [-Wformat=]
         LV_LOG_INFO("Using %d pages to make the roller look infinite", roller->inf_page_cnt);
                                                                                                         ^

+++ src/core/lv_refr.c
@@ -1236,7 +1236,8 @@ static void draw_buf_flush(lv_disp_t * disp)
 
 static void call_flush_cb(lv_disp_drv_t * drv, const lv_area_t * area, lv_color_t * color_p)
 {
-    REFR_TRACE("Calling flush_cb on (%d;%d)(%d;%d) area with %p image pointer", area->x1, area->y1, area->x2, area->y2,
+    REFR_TRACE("Calling flush_cb on (%d;%d)(%d;%d) area with %p image pointer",
+               (int)area->x1, (int)area->y1, (int)area->x2, (int)area->y2,
                (void *)color_p);
 
     lv_area_t offset_area = {
